+++ src/components/reviews/reviews.jsx
@@ -4,6 +4,7 @@ import Form from '../form/form';
 import Comment from '../comment/comment';
 import {useSelector} from 'react-redux';
 import {AuthorizationStatus} from '../../utils/utils';
+import {STATE_SELECTOR} from '../../utils/state-selector';
 
 const getCommentContainer = (comments) => {
   if (comments) {
@@ -17,8 +18,8 @@ const getCommentContainer = (comments) => {
 
 const Reviews = ({id}) => {
 
-  const {authorizationStatus} = useSelector((state) => state.USER);
-  const {comments} = useSelector((state) => state.DATA);
+  const {authorizationStatus} = useSelector(STATE_SELECTOR.USER);
+  const {comments} = useSelector(STATE_SELECTOR.DATA);
 
   return <section className="property__reviews reviews">
     <h2 className="reviews__title">Reviews · <span className="reviews__amount">{comments.length}</span></h2>
