+++ src/utils/state-selector.js
@@ -1,0 +1,5 @@
+export const STATE_SELECTOR = {
+  DATA: (state) => state.DATA,
+  USER: (state) => state.USER,
+  MAIN: (state) => state.MAIN
+};
